Add vectorize config to tables (#1489)
