+++ src/main/java/io/stargate/sgv2/jsonapi/service/resolver/CreateTableCommandResolver.java
@@ -1,8 +1,12 @@
 package io.stargate.sgv2.jsonapi.service.resolver;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import io.stargate.sgv2.jsonapi.api.model.command.CommandContext;
 import io.stargate.sgv2.jsonapi.api.model.command.impl.CreateTableCommand;
+import io.stargate.sgv2.jsonapi.api.model.command.impl.VectorizeConfig;
 import io.stargate.sgv2.jsonapi.api.model.command.table.definition.PrimaryKey;
+import io.stargate.sgv2.jsonapi.api.model.command.table.definition.datatype.ComplexTypes;
 import io.stargate.sgv2.jsonapi.config.DebugModeConfig;
 import io.stargate.sgv2.jsonapi.config.OperationsConfig;
 import io.stargate.sgv2.jsonapi.exception.SchemaException;
@@ -13,13 +17,18 @@ import io.stargate.sgv2.jsonapi.service.operation.tables.CreateTableAttemptBuild
 import io.stargate.sgv2.jsonapi.service.operation.tables.KeyspaceDriverExceptionHandler;
 import io.stargate.sgv2.jsonapi.service.schema.tables.ApiDataType;
 import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
 @ApplicationScoped
 public class CreateTableCommandResolver implements CommandResolver<CreateTableCommand> {
+  @Inject ObjectMapper objectMapper;
+  @Inject VectorizeConfigValidator validateVectorize;
+
   @Override
   public Operation resolveKeyspaceCommand(
       CommandContext<KeyspaceSchemaObject> ctx, CreateTableCommand command) {
@@ -29,6 +38,22 @@ public class CreateTableCommandResolver implements CommandResolver<CreateTableCo
         command.definition().columns().entrySet().stream()
             .collect(Collectors.toMap(Map.Entry::getKey, e -> e.getValue().getApiDataType()));
     List<String> partitionKeys = Arrays.stream(command.definition().primaryKey().keys()).toList();
+
+    Map<String, VectorizeConfig> vectorizeConfigMap =
+        command.definition().columns().entrySet().stream()
+            .filter(
+                e ->
+                    e.getValue() instanceof ComplexTypes.VectorType vt
+                        && vt.getVectorConfig() != null)
+            .collect(
+                Collectors.toMap(
+                    Map.Entry::getKey,
+                    e -> {
+                      ComplexTypes.VectorType vectorType = ((ComplexTypes.VectorType) e.getValue());
+                      final VectorizeConfig vectorConfig = vectorType.getVectorConfig();
+                      validateVectorize.validateService(vectorConfig, vectorType.getDimension());
+                      return vectorConfig;
+                    }));
 
     if (partitionKeys.isEmpty()) {
       throw SchemaException.Code.MISSING_PRIMARY_KEYS.get();
@@ -57,7 +82,17 @@ public class CreateTableCommandResolver implements CommandResolver<CreateTableCo
         });
 
     // set to empty will be used when vectorize is  supported
-    String comment = "";
+    Map<String, String> customProperties = new HashMap<>();
+    try {
+      customProperties.put("com.datastax.data-api.schema-type", "table");
+      // Versioning for schema json. This needs can be adapted in future as needed
+      customProperties.put("com.datastax.data-api.schema-def-version", "1");
+      String vectorizeConfigToStore = objectMapper.writeValueAsString(vectorizeConfigMap);
+      customProperties.put("com.datastax.data-api.vectorize-config", vectorizeConfigToStore);
+    } catch (JsonProcessingException e) {
+      // this should never happen
+      throw new RuntimeException(e);
+    }
 
     var attempt =
         new CreateTableAttemptBuilder(0, ctx.schemaObject())
@@ -69,7 +104,7 @@ public class CreateTableCommandResolver implements CommandResolver<CreateTableCo
             .partitionKeys(partitionKeys)
             .clusteringKeys(clusteringKeys)
             .ifNotExists(ifNotExists)
-            .comment(comment)
+            .customProperties(customProperties)
             .build();
     var attempts = new OperationAttemptContainer<>(List.of(attempt));
