+++ src/test/java/io/stargate/sgv2/jsonapi/api/v1/tables/CreateTableIntegrationTest.java
@@ -3,6 +3,7 @@ package io.stargate.sgv2.jsonapi.api.v1.tables;
 import io.quarkus.test.common.WithTestResource;
 import io.quarkus.test.junit.QuarkusIntegrationTest;
 import io.stargate.sgv2.jsonapi.api.model.command.table.definition.datatype.ColumnType;
+import io.stargate.sgv2.jsonapi.exception.ErrorCodeV1;
 import io.stargate.sgv2.jsonapi.exception.SchemaException;
 import io.stargate.sgv2.jsonapi.testresource.DseTestResource;
 import java.util.ArrayList;
@@ -46,50 +47,50 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                  {
-                                     "name": "allTypesTable",
-                                     "definition": {
-                                         "columns": {
-                                             "ascii_type": "ascii",
-                                             "bigint_type": "bigint",
-                                             "blob_type": "blob",
-                                             "boolean_type": "boolean",
-                                             "date_type": "date",
-                                             "decimal_type": "decimal",
-                                             "double_type": "double",
-                                             "duration_type": "duration",
-                                             "float_type": "float",
-                                             "inet_type": "inet",
-                                             "int_type": "int",
-                                             "smallint_type": "smallint",
-                                             "text_type": "text",
-                                             "time_type": "time",
-                                             "timestamp_type": "timestamp",
-                                             "tinyint_type": "tinyint",
-                                             "uuid_type": "uuid",
-                                             "varint_type": "varint",
-                                             "map_type": {
-                                                "type": "map",
-                                                "keyType": "text",
-                                                "valueType": "int"
-                                             },
-                                             "list_type": {
-                                                "type": "list",
-                                                "valueType": "text"
-                                             },
-                                             "set_type": {
-                                                "type": "set",
-                                                "valueType": "text"
-                                             },
-                                             "vector_type": {
-                                                "type": "vector",
-                                                "dimension": 5
-                                             }
-                                         },
-                                         "primaryKey": "text_type"
-                                     }
-                                  }
-                                  """,
+                                                    {
+                                                       "name": "allTypesTable",
+                                                       "definition": {
+                                                           "columns": {
+                                                               "ascii_type": "ascii",
+                                                               "bigint_type": "bigint",
+                                                               "blob_type": "blob",
+                                                               "boolean_type": "boolean",
+                                                               "date_type": "date",
+                                                               "decimal_type": "decimal",
+                                                               "double_type": "double",
+                                                               "duration_type": "duration",
+                                                               "float_type": "float",
+                                                               "inet_type": "inet",
+                                                               "int_type": "int",
+                                                               "smallint_type": "smallint",
+                                                               "text_type": "text",
+                                                               "time_type": "time",
+                                                               "timestamp_type": "timestamp",
+                                                               "tinyint_type": "tinyint",
+                                                               "uuid_type": "uuid",
+                                                               "varint_type": "varint",
+                                                               "map_type": {
+                                                                  "type": "map",
+                                                                  "keyType": "text",
+                                                                  "valueType": "int"
+                                                               },
+                                                               "list_type": {
+                                                                  "type": "list",
+                                                                  "valueType": "text"
+                                                               },
+                                                               "set_type": {
+                                                                  "type": "set",
+                                                                  "valueType": "text"
+                                                               },
+                                                               "vector_type": {
+                                                                  "type": "vector",
+                                                                  "dimension": 5
+                                                               }
+                                                           },
+                                                           "primaryKey": "text_type"
+                                                       }
+                                                    }
+                                                    """,
                   "allTypesTable",
                   false,
                   null,
@@ -99,24 +100,24 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                                {
-                                                   "name": "primaryKeyAsStringTable",
-                                                   "definition": {
-                                                       "columns": {
-                                                           "id": {
-                                                               "type": "text"
-                                                           },
-                                                           "age": {
-                                                               "type": "int"
-                                                           },
-                                                           "name": {
-                                                               "type": "text"
-                                                           }
-                                                       },
-                                                       "primaryKey": "id"
-                                                   }
-                                                }
-                                                """,
+                                                                  {
+                                                                     "name": "primaryKeyAsStringTable",
+                                                                     "definition": {
+                                                                         "columns": {
+                                                                             "id": {
+                                                                                 "type": "text"
+                                                                             },
+                                                                             "age": {
+                                                                                 "type": "int"
+                                                                             },
+                                                                             "name": {
+                                                                                 "type": "text"
+                                                                             }
+                                                                         },
+                                                                         "primaryKey": "id"
+                                                                     }
+                                                                  }
+                                                                  """,
                   "primaryKeyAsStringTable",
                   false,
                   null,
@@ -126,24 +127,24 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                    {
-                                       "name": "primaryKeyAsStringTable",
-                                       "definition": {
-                                           "columns": {
-                                               "id": {
-                                                   "type": "text"
-                                               },
-                                               "age": {
-                                                   "type": "int"
-                                               },
-                                               "name": {
-                                                   "type": "text"
-                                               }
-                                           },
-                                           "primaryKey": "id"
-                                       }
-                                    }
-                                    """,
+                                                      {
+                                                         "name": "primaryKeyAsStringTable",
+                                                         "definition": {
+                                                             "columns": {
+                                                                 "id": {
+                                                                     "type": "text"
+                                                                 },
+                                                                 "age": {
+                                                                     "type": "int"
+                                                                 },
+                                                                 "name": {
+                                                                     "type": "text"
+                                                                 }
+                                                             },
+                                                             "primaryKey": "id"
+                                                         }
+                                                      }
+                                                      """,
                   "primaryKeyAsStringTable",
                   false,
                   null,
@@ -154,21 +155,21 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                    {
-                                          "name": "primaryKeyWithQuotable",
-                                          "definition": {
-                                            "primaryKey": "_id",
-                                            "columns": {
-                                              "_id": {
-                                                "type": "text"
-                                              },
-                                              "name": {
-                                                "type": "text"
-                                              }
-                                            }
-                                          }
-                                      }
-                                    """,
+                                                      {
+                                                            "name": "primaryKeyWithQuotable",
+                                                            "definition": {
+                                                              "primaryKey": "_id",
+                                                              "columns": {
+                                                                "_id": {
+                                                                  "type": "text"
+                                                                },
+                                                                "name": {
+                                                                  "type": "text"
+                                                                }
+                                                              }
+                                                            }
+                                                        }
+                                                      """,
                   "primaryKeyWithQuotable",
                   false,
                   null,
@@ -179,18 +180,18 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                    {
-                                       "name": "columnTypeusingShortHandTable",
-                                       "definition": {
-                                           "columns": {
-                                               "id": "text",
-                                               "age": "int",
-                                               "name": "text"
-                                           },
-                                           "primaryKey": "id"
-                                       }
-                                    }
-                                    """,
+                                                      {
+                                                         "name": "columnTypeusingShortHandTable",
+                                                         "definition": {
+                                                             "columns": {
+                                                                 "id": "text",
+                                                                 "age": "int",
+                                                                 "name": "text"
+                                                             },
+                                                             "primaryKey": "id"
+                                                         }
+                                                      }
+                                                      """,
                   "columnTypeusingShortHandTable",
                   false,
                   null,
@@ -201,31 +202,31 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                                {
-                                                  "name": "primaryKeyAsJsonObjectTable",
-                                                  "definition": {
-                                                    "columns": {
-                                                      "id": {
-                                                        "type": "text"
-                                                      },
-                                                      "age": {
-                                                        "type": "int"
-                                                      },
-                                                      "name": {
-                                                        "type": "text"
-                                                      }
-                                                    },
-                                                    "primaryKey": {
-                                                      "partitionBy": [
-                                                        "id"
-                                                      ],
-                                                      "partitionSort" : {
-                                                        "name" : 1, "age" : -1
-                                                      }
-                                                    }
-                                                  }
-                                                }
-                                                """,
+                                                                  {
+                                                                    "name": "primaryKeyAsJsonObjectTable",
+                                                                    "definition": {
+                                                                      "columns": {
+                                                                        "id": {
+                                                                          "type": "text"
+                                                                        },
+                                                                        "age": {
+                                                                          "type": "int"
+                                                                        },
+                                                                        "name": {
+                                                                          "type": "text"
+                                                                        }
+                                                                      },
+                                                                      "primaryKey": {
+                                                                        "partitionBy": [
+                                                                          "id"
+                                                                        ],
+                                                                        "partitionSort" : {
+                                                                          "name" : 1, "age" : -1
+                                                                        }
+                                                                      }
+                                                                    }
+                                                                  }
+                                                                  """,
                   "primaryKeyAsJsonObjectTable",
                   false,
                   null,
@@ -238,24 +239,24 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                    {
-                                     "name": "invalidPrimaryKeyTable",
-                                     "definition": {
-                                         "columns": {
-                                             "id": {
-                                                 "type": "text"
-                                             },
-                                             "age": {
-                                                 "type": "int"
-                                             },
-                                             "name": {
-                                                 "type": "text"
-                                             }
-                                         },
-                                         "primaryKey": "error_column"
-                                     }
-                                    }
-                                    """,
+                                                      {
+                                                       "name": "invalidPrimaryKeyTable",
+                                                       "definition": {
+                                                           "columns": {
+                                                               "id": {
+                                                                   "type": "text"
+                                                               },
+                                                               "age": {
+                                                                   "type": "int"
+                                                               },
+                                                               "name": {
+                                                                   "type": "text"
+                                                               }
+                                                           },
+                                                           "primaryKey": "error_column"
+                                                       }
+                                                      }
+                                                      """,
                   "invalidPrimaryKeyTable",
                   true,
                   missingDefinition.code,
@@ -266,25 +267,25 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                    {
-                                      "name": "invalidPartitionByTable",
-                                      "definition": {
-                                        "columns": {
-                                          "id": "text",
-                                          "age": "int",
-                                          "name": "text"
-                                        },
-                                        "primaryKey": {
-                                          "partitionBy": [
-                                            "error_column"
-                                          ],
-                                          "partitionSort" : {
-                                            "name" : 1, "age" : -1
-                                          }
-                                        }
-                                      }
-                                    }
-                                    """,
+                                                      {
+                                                        "name": "invalidPartitionByTable",
+                                                        "definition": {
+                                                          "columns": {
+                                                            "id": "text",
+                                                            "age": "int",
+                                                            "name": "text"
+                                                          },
+                                                          "primaryKey": {
+                                                            "partitionBy": [
+                                                              "error_column"
+                                                            ],
+                                                            "partitionSort" : {
+                                                              "name" : 1, "age" : -1
+                                                            }
+                                                          }
+                                                        }
+                                                      }
+                                                      """,
                   "invalidPartitionByTable",
                   true,
                   missingDefinition.code,
@@ -295,25 +296,25 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                    {
-                                        "name": "invalidPartitionSortTable",
-                                        "definition": {
-                                          "columns": {
-                                            "id": "text",
-                                            "age": "int",
-                                            "name": "text"
-                                          },
-                                          "primaryKey": {
-                                            "partitionBy": [
-                                              "id"
-                                            ],
-                                            "partitionSort" : {
-                                              "error_column" : 1, "age" : -1
-                                            }
-                                          }
-                                        }
-                                      }
-                                    """,
+                                                      {
+                                                          "name": "invalidPartitionSortTable",
+                                                          "definition": {
+                                                            "columns": {
+                                                              "id": "text",
+                                                              "age": "int",
+                                                              "name": "text"
+                                                            },
+                                                            "primaryKey": {
+                                                              "partitionBy": [
+                                                                "id"
+                                                              ],
+                                                              "partitionSort" : {
+                                                                "error_column" : 1, "age" : -1
+                                                              }
+                                                            }
+                                                          }
+                                                        }
+                                                      """,
                   "invalidPartitionSortTable",
                   true,
                   missingDefinition.code,
@@ -325,25 +326,25 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                    {
-                                      "name": "invalidPartitionSortOrderingValueTable",
-                                      "definition": {
-                                        "columns": {
-                                          "id": "text",
-                                          "age": "int",
-                                          "name": "text"
-                                        },
-                                        "primaryKey": {
-                                          "partitionBy": [
-                                            "id"
-                                          ],
-                                          "partitionSort" : {
-                                            "id" : 1, "age" : 0
-                                          }
-                                        }
-                                      }
-                                    }
-                                    """,
+                                                      {
+                                                        "name": "invalidPartitionSortOrderingValueTable",
+                                                        "definition": {
+                                                          "columns": {
+                                                            "id": "text",
+                                                            "age": "int",
+                                                            "name": "text"
+                                                          },
+                                                          "primaryKey": {
+                                                            "partitionBy": [
+                                                              "id"
+                                                            ],
+                                                            "partitionSort" : {
+                                                              "id" : 1, "age" : 0
+                                                            }
+                                                          }
+                                                        }
+                                                      }
+                                                      """,
                   "invalidPartitionSortOrderingValueTable",
                   true,
                   se.code,
@@ -354,25 +355,25 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                    {
-                                        "name": "invalidPartitionSortOrderingValueTypeTable",
-                                        "definition": {
-                                          "columns": {
-                                            "id": "text",
-                                            "age": "int",
-                                            "name": "text"
-                                          },
-                                          "primaryKey": {
-                                            "partitionBy": [
-                                              "id"
-                                            ],
-                                            "partitionSort" : {
-                                              "id" : 1, "age" : "invalid"
-                                            }
-                                          }
-                                        }
-                                      }
-                                    """,
+                                                      {
+                                                          "name": "invalidPartitionSortOrderingValueTypeTable",
+                                                          "definition": {
+                                                            "columns": {
+                                                              "id": "text",
+                                                              "age": "int",
+                                                              "name": "text"
+                                                            },
+                                                            "primaryKey": {
+                                                              "partitionBy": [
+                                                                "id"
+                                                              ],
+                                                              "partitionSort" : {
+                                                                "id" : 1, "age" : "invalid"
+                                                              }
+                                                            }
+                                                          }
+                                                        }
+                                                      """,
                   "invalidPartitionSortOrderingValueTypeTable",
                   true,
                   se.code,
@@ -391,26 +392,26 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                    {
-                                      "name": "invalidColumnTypeTable",
-                                      "definition": {
-                                        "columns": {
-                                          "id": "invalid_type",
-                                          "age": "int",
-                                          "name": "text"
-                                        },
-                                        "primaryKey": {
-                                          "partitionBy": [
-                                              "id"
-                                          ],
-                                          "partitionSort": {
-                                              "id": 1,
-                                              "age": -1
-                                          }
-                                        }
-                                      }
-                                    }
-                                    """,
+                                                      {
+                                                        "name": "invalidColumnTypeTable",
+                                                        "definition": {
+                                                          "columns": {
+                                                            "id": "invalid_type",
+                                                            "age": "int",
+                                                            "name": "text"
+                                                          },
+                                                          "primaryKey": {
+                                                            "partitionBy": [
+                                                                "id"
+                                                            ],
+                                                            "partitionSort": {
+                                                                "id": 1,
+                                                                "age": -1
+                                                            }
+                                                          }
+                                                        }
+                                                      }
+                                                      """,
                   "invalidColumnTypeTable",
                   true,
                   invalidType.code,
@@ -422,26 +423,26 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                                {
-                                                  "name": "invalidColumnTypeTable",
-                                                  "definition": {
-                                                    "columns": {
-                                                      "id": null,
-                                                      "age": "int",
-                                                      "name": "text"
-                                                    },
-                                                    "primaryKey": {
-                                                      "partitionBy": [
-                                                          "id"
-                                                      ],
-                                                      "partitionSort": {
-                                                          "id": 1,
-                                                          "age": -1
-                                                      }
-                                                    }
-                                                  }
-                                                }
-                                                """,
+                                                                  {
+                                                                    "name": "invalidColumnTypeTable",
+                                                                    "definition": {
+                                                                      "columns": {
+                                                                        "id": null,
+                                                                        "age": "int",
+                                                                        "name": "text"
+                                                                      },
+                                                                      "primaryKey": {
+                                                                        "partitionBy": [
+                                                                            "id"
+                                                                        ],
+                                                                        "partitionSort": {
+                                                                            "id": 1,
+                                                                            "age": -1
+                                                                        }
+                                                                      }
+                                                                    }
+                                                                  }
+                                                                  """,
                   "invalidColumnTypeTable",
                   true,
                   columnTypeInvalid.code,
@@ -453,22 +454,22 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                        {
-                                          "name": "invalidMapType",
-                                          "definition": {
-                                            "columns": {
-                                              "id": "text",
-                                              "age": "int",
-                                              "name": "text",
-                                              "map_type": {
-                                                "type": "map",
-                                                "keyType": "text"
-                                              }
-                                            },
-                                            "primaryKey": "id"
-                                          }
-                                        }
-                                                            """,
+                                                          {
+                                                            "name": "invalidMapType",
+                                                            "definition": {
+                                                              "columns": {
+                                                                "id": "text",
+                                                                "age": "int",
+                                                                "name": "text",
+                                                                "map_type": {
+                                                                  "type": "map",
+                                                                  "keyType": "text"
+                                                                }
+                                                              },
+                                                              "primaryKey": "id"
+                                                            }
+                                                          }
+                                                                              """,
                   "invalidMapType value type not provided",
                   true,
                   invalidMapType.code,
@@ -478,22 +479,22 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                        {
-                                          "name": "invalidMapType",
-                                          "definition": {
-                                            "columns": {
-                                              "id": "text",
-                                              "age": "int",
-                                              "name": "text",
-                                              "map_type": {
-                                                "type": "map",
-                                                "valueType": "text"
-                                              }
-                                            },
-                                            "primaryKey": "id"
-                                          }
-                                        }
-                                                            """,
+                                                          {
+                                                            "name": "invalidMapType",
+                                                            "definition": {
+                                                              "columns": {
+                                                                "id": "text",
+                                                                "age": "int",
+                                                                "name": "text",
+                                                                "map_type": {
+                                                                  "type": "map",
+                                                                  "valueType": "text"
+                                                                }
+                                                              },
+                                                              "primaryKey": "id"
+                                                            }
+                                                          }
+                                                                              """,
                   "invalidMapType key type not provided",
                   true,
                   invalidMapType.code,
@@ -503,23 +504,23 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                        {
-                                          "name": "invalidMapType",
-                                          "definition": {
-                                            "columns": {
-                                              "id": "text",
-                                              "age": "int",
-                                              "name": "text",
-                                              "map_type": {
-                                                "type": "map",
-                                                "valueType": "list",
-                                                "keyType": "text"
-                                              }
-                                            },
-                                            "primaryKey": "id"
-                                          }
-                                        }
-                                                            """,
+                                                          {
+                                                            "name": "invalidMapType",
+                                                            "definition": {
+                                                              "columns": {
+                                                                "id": "text",
+                                                                "age": "int",
+                                                                "name": "text",
+                                                                "map_type": {
+                                                                  "type": "map",
+                                                                  "valueType": "list",
+                                                                  "keyType": "text"
+                                                                }
+                                                              },
+                                                              "primaryKey": "id"
+                                                            }
+                                                          }
+                                                                              """,
                   "invalidMapType not primitive type provided",
                   true,
                   invalidMapType.code,
@@ -529,23 +530,23 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                        {
-                                          "name": "invalidMapType",
-                                          "definition": {
-                                            "columns": {
-                                              "id": "text",
-                                              "age": "int",
-                                              "name": "text",
-                                              "map_type": {
-                                                "type": "map",
-                                                "valueType": "text",
-                                                "keyType": "list"
-                                              }
-                                            },
-                                            "primaryKey": "id"
-                                          }
-                                        }
-                                                            """,
+                                                          {
+                                                            "name": "invalidMapType",
+                                                            "definition": {
+                                                              "columns": {
+                                                                "id": "text",
+                                                                "age": "int",
+                                                                "name": "text",
+                                                                "map_type": {
+                                                                  "type": "map",
+                                                                  "valueType": "text",
+                                                                  "keyType": "list"
+                                                                }
+                                                              },
+                                                              "primaryKey": "id"
+                                                            }
+                                                          }
+                                                                              """,
                   "invalidMapType not primitive type provided",
                   true,
                   invalidMapType.code,
@@ -557,21 +558,21 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                        {
-                                          "name": "invalidListType",
-                                          "definition": {
-                                            "columns": {
-                                              "id": "text",
-                                              "age": "int",
-                                              "name": "text",
-                                              "list_type": {
-                                                "type": "list"
-                                              }
-                                            },
-                                            "primaryKey": "id"
-                                          }
-                                        }
-                                                            """,
+                                                          {
+                                                            "name": "invalidListType",
+                                                            "definition": {
+                                                              "columns": {
+                                                                "id": "text",
+                                                                "age": "int",
+                                                                "name": "text",
+                                                                "list_type": {
+                                                                  "type": "list"
+                                                                }
+                                                              },
+                                                              "primaryKey": "id"
+                                                            }
+                                                          }
+                                                                              """,
                   "invalidListType value type not provided",
                   true,
                   invalidListType.code,
@@ -581,22 +582,22 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                        {
-                                          "name": "invalidListType",
-                                          "definition": {
-                                            "columns": {
-                                              "id": "text",
-                                              "age": "int",
-                                              "name": "text",
-                                              "list_type": {
-                                                "type": "list",
-                                                "valueType": "list"
-                                              }
-                                            },
-                                            "primaryKey": "id"
-                                          }
-                                        }
-                                                            """,
+                                                          {
+                                                            "name": "invalidListType",
+                                                            "definition": {
+                                                              "columns": {
+                                                                "id": "text",
+                                                                "age": "int",
+                                                                "name": "text",
+                                                                "list_type": {
+                                                                  "type": "list",
+                                                                  "valueType": "list"
+                                                                }
+                                                              },
+                                                              "primaryKey": "id"
+                                                            }
+                                                          }
+                                                                              """,
                   "invalidListType not primitive type provided",
                   true,
                   invalidListType.code,
@@ -608,21 +609,21 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                        {
-                                          "name": "invalidSetType",
-                                          "definition": {
-                                            "columns": {
-                                              "id": "text",
-                                              "age": "int",
-                                              "name": "text",
-                                              "set_type": {
-                                                "type": "set"
-                                              }
-                                            },
-                                            "primaryKey": "id"
-                                          }
-                                        }
-                                                            """,
+                                                          {
+                                                            "name": "invalidSetType",
+                                                            "definition": {
+                                                              "columns": {
+                                                                "id": "text",
+                                                                "age": "int",
+                                                                "name": "text",
+                                                                "set_type": {
+                                                                  "type": "set"
+                                                                }
+                                                              },
+                                                              "primaryKey": "id"
+                                                            }
+                                                          }
+                                                                              """,
                   "invalidSetType value type not provided",
                   true,
                   invalidSetType.code,
@@ -632,22 +633,22 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                        {
-                                          "name": "invalidSetType",
-                                          "definition": {
-                                            "columns": {
-                                              "id": "text",
-                                              "age": "int",
-                                              "name": "text",
-                                              "set_type": {
-                                                "type": "set",
-                                                "valueType": "list"
-                                              }
-                                            },
-                                            "primaryKey": "id"
-                                          }
-                                        }
-                                                            """,
+                                                          {
+                                                            "name": "invalidSetType",
+                                                            "definition": {
+                                                              "columns": {
+                                                                "id": "text",
+                                                                "age": "int",
+                                                                "name": "text",
+                                                                "set_type": {
+                                                                  "type": "set",
+                                                                  "valueType": "list"
+                                                                }
+                                                              },
+                                                              "primaryKey": "id"
+                                                            }
+                                                          }
+                                                                              """,
                   "invalidSetType not primitive type provided",
                   true,
                   invalidSetType.code,
@@ -660,22 +661,22 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                        {
-                                          "name": "invalidVectorType",
-                                          "definition": {
-                                            "columns": {
-                                              "id": "text",
-                                              "age": "int",
-                                              "name": "text",
-                                              "vector_type": {
-                                                "type": "vector",
-                                                "dimension": -5
-                                              }
-                                            },
-                                            "primaryKey": "id"
-                                          }
-                                        }
-                                                            """,
+                                                          {
+                                                            "name": "invalidVectorType",
+                                                            "definition": {
+                                                              "columns": {
+                                                                "id": "text",
+                                                                "age": "int",
+                                                                "name": "text",
+                                                                "vector_type": {
+                                                                  "type": "vector",
+                                                                  "dimension": -5
+                                                                }
+                                                              },
+                                                              "primaryKey": "id"
+                                                            }
+                                                          }
+                                                                              """,
                   "invalidVectorType value type not provided",
                   true,
                   invalidVectorType.code,
@@ -685,27 +686,157 @@ class CreateTableIntegrationTest extends AbstractTableIntegrationTestBase {
           Arguments.of(
               new CreateTableTestData(
                   """
-                                        {
-                                          "name": "invalidVectorType",
-                                          "definition": {
-                                            "columns": {
-                                              "id": "text",
-                                              "age": "int",
-                                              "name": "text",
-                                              "vector_type": {
-                                                "type": "vector",
-                                                "dimension": "aaa"
-                                              }
-                                            },
-                                            "primaryKey": "id"
-                                          }
-                                        }
-                                                            """,
+                                                          {
+                                                            "name": "invalidVectorType",
+                                                            "definition": {
+                                                              "columns": {
+                                                                "id": "text",
+                                                                "age": "int",
+                                                                "name": "text",
+                                                                "vector_type": {
+                                                                  "type": "vector",
+                                                                  "dimension": "aaa"
+                                                                }
+                                                              },
+                                                              "primaryKey": "id"
+                                                            }
+                                                          }
+                                                                              """,
                   "invalidVectorType not primitive type provided",
                   true,
                   invalidVectorType.code,
                   invalidVectorType.body)));
+      // vector type with vectorize
+      testCases.add(
+          Arguments.of(
+              new CreateTableTestData(
+                  """
+                                    {
+                                       "name": "vectorizeConfigTest",
+                                       "definition": {
+                                           "columns": {
+                                               "id": {
+                                                   "type": "text"
+                                               },
+                                               "age": {
+                                                   "type": "int"
+                                               },
+                                               "content": {
+                                                 "type": "vector",
+                                                 "dimension": 1024,
+                                                 "service": {
+                                                   "provider": "nvidia",
+                                                   "modelName": "NV-Embed-QA"
+                                                 }
+                                               }
+                                           },
+                                           "primaryKey": "id"
+                                       }
+                                    }
+                                    """,
+                  "primaryKeyAsStringTable",
+                  false,
+                  null,
+                  null)));
 
+      // vector type with invalid vectorixe config
+      testCases.add(
+          Arguments.of(
+              new CreateTableTestData(
+                  """
+                  {
+                     "name": "invalidVectorizeServiceNameConfig",
+                     "definition": {
+                         "columns": {
+                             "id": {
+                                 "type": "text"
+                             },
+                             "age": {
+                                 "type": "int"
+                             },
+                             "content": {
+                               "type": "vector",
+                               "dimension": 1024,
+                               "service": {
+                                 "provider": "invalid_service",
+                                 "modelName": "NV-Embed-QA"
+                               }
+                             }
+                         },
+                         "primaryKey": "id"
+                     }
+                  }
+                  """,
+                  "invalidVectorizeServiceNameConfig",
+                  true,
+                  ErrorCodeV1.INVALID_CREATE_COLLECTION_OPTIONS.name(),
+                  "The provided options are invalid: Service provider 'invalid_service' is not supported")));
+
+      // vector type with invalid model name config
+      testCases.add(
+          Arguments.of(
+              new CreateTableTestData(
+                  """
+            {
+               "name": "invalidVectorizeModelNameConfig",
+               "definition": {
+                   "columns": {
+                       "id": {
+                           "type": "text"
+                       },
+                       "age": {
+                           "type": "int"
+                       },
+                       "content": {
+                         "type": "vector",
+                         "dimension": 1024,
+                         "service": {
+                          "provider": "mistral",
+                          "modelName": "mistral-embed-invalid"
+                        }
+                       }
+                   },
+                   "primaryKey": "id"
+               }
+            }
+            """,
+                  "invalidVectorizeModelNameConfig",
+                  true,
+                  ErrorCodeV1.INVALID_CREATE_COLLECTION_OPTIONS.name(),
+                  "The provided options are invalid: Model name 'mistral-embed-invalid' for provider 'mistral' is not supported")));
+
+      // vector type with dimension mismatch
+      testCases.add(
+          Arguments.of(
+              new CreateTableTestData(
+                  """
+            {
+               "name": "invalidVectorizeModelNameConfig",
+               "definition": {
+                   "columns": {
+                       "id": {
+                           "type": "text"
+                       },
+                       "age": {
+                           "type": "int"
+                       },
+                       "content": {
+                         "type": "vector",
+                         "dimension": 1536,
+                         "service": {
+                          "provider": "mistral",
+                          "modelName": "mistral-embed"
+                        }
+                       }
+                   },
+                   "primaryKey": "id"
+               }
+            }
+            """,
+                  "invalidVectorizeModelNameConfig",
+                  true,
+                  ErrorCodeV1.INVALID_CREATE_COLLECTION_OPTIONS.name(),
+                  "The provided options are invalid: The provided dimension value '1536' doesn't match the model's supported dimension value '1024'")));
       return testCases.stream();
     }
   }
